requirements.txt
Flask==2.1.0
spotipy==2.19.0
google~=3.0.0
requests~=2.27.1
Pillow~=9.0.0